Obtain the stats for a group of samples

This function takes a forward and reverse bam files for a group of samples and returns alignment stats.

get_stats(bam_dir, sample_names, output_dir = "stats", time = FALSE)

Arguments

bam_dir

a file path to a directory where the bam files associated with the sample names will be stored.

sample_names

a character vector that indicates the name of the samples as they are listed in the file names.

output_dir

a file path to where you'd like the stats to be saved to. By default its "stats" folder. This folder will be created if it doesn't exist.

time

TRUE/FALSE you want the duration this takes to run printed out

Examples

if (FALSE) { # \dontrun{

# This will be the directory that contains all the file path
bam_dir <- file.path(example_data_folder(), "bam")

# These MUST be names that are listed in the bam file name itself.
# There needs to be exactly 2 bam files per sample
sample_names <- c("sample1", "sample2", "sample3")

stats_df <- get_stats(
  bam_dir = bam_dir,
  sample_names = sample_names,
  output_dir = file.path(example_data_folder(), "bam_stats")
)

# We can write these to CSV files like this
} # }